// web/src/api/generated/models/PrivateNoteDetail.ts
export interface PrivateNoteDetail {
    /**
     * 
     * @type {string}
     * @memberof PrivateNoteDetail
     */
    body: string;
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof PrivateNoteDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof PrivateNoteDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof PrivateNoteDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof PrivateNoteDetail
     */
    ownerId: string;
    /**
     * 
     * @type {boolean}
     * @memberof PrivateNoteDetail
     */
    pinned: boolean;
    /**
     * 
     * @type {string}
     * @memberof PrivateNoteDetail
     */
    spaceId: string;
    /**
     * 
     * @type {string}
     * @memberof PrivateNoteDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof PrivateNoteDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof PrivateNoteDetail
     */
    version: number;
}